chore: Update Cryptography Dependencies (#1195)
