+++ Cargo.toml
@@ -9,7 +9,7 @@ exclude = [
 members = [
     "crates/dwallet-mpc-centralized-party",
     "crates/dwallet-mpc-types",
-    "crates/shared-wasm-class-groups",
+    "crates/message-digest",
     "crates/dwallet-classgroups-types",
     "crates/ika",
     "crates/ika-config",
@@ -74,12 +74,13 @@ unexpected_cfgs = { level = "warn", check-cfg = [
 
 # Dependencies that should be kept in sync through the whole workspace
 [workspace.dependencies]
-k256 = { git = "https://github.com/RustCrypto/elliptic-curves.git", tag = "k256/v0.14.0-pre.2", features = ["arithmetic", "critical-section", "precomputed-tables", "serde", "ecdsa", "hash2curve", "alloc"], default-features = false }
-mpc = { git = "https://github.com/dwallet-labs/cryptography-private", rev = "e5eaa62"}
-class_groups = { git = "https://github.com/dwallet-labs/cryptography-private", rev = "e5eaa62", features = ["threshold", "parallel"] }
-twopc_mpc = { git = "https://github.com/dwallet-labs/cryptography-private", features = ["secp256k1", "bulletproofs", "paillier", "class_groups"], rev = "e5eaa62"}
-group = { git = "https://github.com/dwallet-labs/cryptography-private", rev = "e5eaa62"}
-homomorphic_encryption = { git = "https://github.com/dwallet-labs/cryptography-private", rev = "e5eaa62"}
+k256 = { version = "0.14.0-pre.7", features = ["arithmetic", "critical-section", "precomputed-tables", "serde", "ecdsa", "hash2curve", "alloc"], default-features = false }
+mpc = { git = "https://github.com/dwallet-labs/cryptography-private", rev = "7dd2b76"}
+class_groups = { git = "https://github.com/dwallet-labs/cryptography-private", rev = "7dd2b76", features = ["threshold"] }
+commitment = { git = "https://github.com/dwallet-labs/cryptography-private", rev = "7dd2b76" }
+twopc_mpc = { git = "https://github.com/dwallet-labs/cryptography-private", features = ["secp256k1", "class_groups"], rev = "7dd2b76"}
+group = { git = "https://github.com/dwallet-labs/cryptography-private", features = ["os_rng"], rev = "7dd2b76"}
+homomorphic_encryption = { git = "https://github.com/dwallet-labs/cryptography-private", rev = "7dd2b76"}
 anyhow = "1.0.71"
 arrow = "52"
 arrow-array = "52"
@@ -257,7 +258,8 @@ prost-types = "0.13.1"
 protobuf = { version = "2.28", features = ["with-bytes"] }
 quinn-proto = "0.11.7"
 quote = "1.0.23"
-rand = "0.8.5"
+rand = "0.9"
+rand_chacha = "0.9"
 rayon = "1.5.3"
 rcgen = "0.13"
 regex = "1.7.1"
@@ -301,7 +303,7 @@ serde_with = "3.8"
 serde_yaml = "0.8.26"
 shell-words = "1.1.0"
 shellexpand = "3.1.0"
-signature = "2.2.0"
+signature = "3.0.0-rc.1"
 similar = "2.4.0"
 simple-server-timing-header = "0.1.1"
 slip10_ed25519 = "0.1.3"
@@ -521,7 +523,7 @@ sui-execution = { git = "https://github.com/MystenLabs/sui", rev = "3b1d6b3bd63f
 consensus-config = { git = "https://github.com/MystenLabs/sui", rev = "3b1d6b3bd63f175b774da557f89af3619b74d783" }
 consensus-core = { git = "https://github.com/MystenLabs/sui", rev = "3b1d6b3bd63f175b774da557f89af3619b74d783" }
 
-crypto-bigint = { git = 'https://github.com/erik-3milabs/crypto-bigint.git', rev = 'a4da280' }
+crypto-bigint = { git = 'https://github.com/erik-3milabs/crypto-bigint.git', rev = "640aacb" }
 
 ### Workspace Members ###
 dwallet-mpc-centralized-party = { path = "crates/dwallet-mpc-centralized-party" }
@@ -555,5 +557,5 @@ spinners = "4.1.0"
 include_dir = "0.7.3"
 
 [patch.crates-io]
-ecdsa = { git = "https://github.com/yaelAbergel/signatures", rev = "7c8523182b1a28e5bcc6bb8d1508fbde1ea845c4"}
-crypto-bigint = { git = 'https://github.com/erik-3milabs/crypto-bigint.git', rev = 'a4da280' }
+curve25519-dalek = { git = 'https://github.com/dalek-cryptography/curve25519-dalek', rev = "b7e4fc0" }
+crypto-bigint = { git = 'https://github.com/erik-3milabs/crypto-bigint.git', rev = "640aacb" }
