+++ crates/dwallet-mpc-centralized-party/Cargo.toml
@@ -6,8 +6,8 @@ version.workspace = true
 [dependencies]
 mpc.workspace = true
 twopc_mpc.workspace = true
-commitment = { git = "https://github.com/dwallet-labs/cryptography-private", rev = "e5eaa62" }
-shared_wasm_class_groups = { path = "../shared-wasm-class-groups" }
+commitment = { git = "https://github.com/dwallet-labs/cryptography-private", rev = "7dd2b76" }
+message-digest = { path = "../message-digest" }
 class_groups.workspace = true
 group.workspace = true
 homomorphic_encryption.workspace = true
@@ -15,10 +15,14 @@ dwallet-mpc-types.workspace = true
 schemars.workspace = true
 anyhow.workspace = true
 bcs.workspace = true
-rand_core = "0.6.4"
-rand_chacha = "0.3.1"
-serde = { version = "^1.0.219", features = ["derive"] }
+rand_core = { version = "0.9", default-features = false }
+rand_chacha = { version = "0.9", default-features = false }
+getrandom = { version = "0.2.16", features = ["js"], optional = true } # TODO: idk why, but we need this, also this is old version
+serde = { version = "1.0", features = ["derive"] }
 crypto-bigint.workspace = true
+
+[features]
+wasm_js = ["group/wasm_js", "dep:getrandom"]
 
 [lints]
 workspace = true
