+++ crates/ika-archival/Cargo.toml
@@ -15,7 +15,7 @@ tracing.workspace = true
 bytes.workspace = true
 num_enum.workspace = true
 futures.workspace = true
-rand.workspace = true
+rand = "0.8" # use old version for compatability with fastcrypto
 object_store.workspace = true
 prometheus.workspace = true
 ika-config.workspace = true
