+++ crates/ika-core/Cargo.toml
@@ -15,12 +15,12 @@ group.workspace = true
 mpc.workspace = true
 twopc_mpc.workspace = true
 homomorphic_encryption.workspace = true
-class_groups.workspace = true
+class_groups = { workspace = true, features = ["threshold", "parallel"]}
 dwallet-classgroups-types.workspace = true
 schemars.workspace = true
-commitment = { git = "https://github.com/dwallet-labs/cryptography-private", rev = "e5eaa62"}
-shared_wasm_class_groups = { path = "../shared-wasm-class-groups" }
-rand_core = "0.6.4"
+commitment.workspace = true
+message-digest = { path = "../message-digest" }
+rand_core = "0.9"
 dwallet-mpc-types.workspace = true
 anyhow = { workspace = true, features = ["backtrace"] }
 arc-swap.workspace = true
@@ -50,6 +50,7 @@ once_cell.workspace = true
 parking_lot.workspace = true
 prometheus.workspace = true
 rand.workspace = true
+rand_chacha.workspace = true
 roaring.workspace = true
 rayon.workspace = true
 reqwest.workspace = true
@@ -110,10 +111,9 @@ num_enum.workspace = true
 strum_macros.workspace = true
 derive_more.workspace = true
 backoff.workspace = true
-sha3 = "0.10.8"
-sha2 = "0.10.8"
+sha3 = "0.11.0-rc.0"
+sha2 = "0.11.0-rc.0"
 base64 = "0.22.1"
-
 
 [dev-dependencies]
 clap.workspace = true
