+++ crates/ika-core/src/dwallet_checkpoints/mod.rs
@@ -21,7 +21,7 @@ use sui_types::base_types::ConciseableName;
 
 use crate::authority::authority_per_epoch_store::AuthorityPerEpochStore;
 
-use crate::dwallet_mpc::MPCSessionLogger;
+use crate::dwallet_mpc::mpc_session::MPCSessionLogger;
 use ika_types::crypto::AuthorityStrongQuorumSignInfo;
 use ika_types::digests::DWalletCheckpointMessageDigest;
 use ika_types::error::{IkaError, IkaResult};
