+++ crates/ika-core/src/dwallet_mpc/dwallet_mpc_service.rs
@@ -7,7 +7,7 @@ use crate::authority::authority_per_epoch_store::AuthorityPerEpochStore;
 use crate::consensus_adapter::SubmitToConsensus;
 use crate::dwallet_mpc::dwallet_mpc_metrics::DWalletMPCMetrics;
 use crate::dwallet_mpc::mpc_manager::{DWalletMPCDBMessage, DWalletMPCManager};
-use crate::dwallet_mpc::session_info_from_event;
+use crate::dwallet_mpc::mpc_session::session_info_from_event;
 use dwallet_mpc_types::dwallet_mpc::{MPCSessionStatus, NetworkDecryptionKeyPublicData};
 use ika_config::NodeConfig;
 use ika_sui_client::SuiConnectorClient;
