+++ crates/ika-core/src/dwallet_mpc/native_computations/encrypt_user_share.rs
@@ -2,11 +2,24 @@ use dwallet_mpc_types::dwallet_mpc::{
     SerializedWrappedMPCPublicOutput, VersionedDwalletDKGSecondRoundPublicOutput,
     VersionedEncryptedUserShare,
 };
+use group::OsCsRng;
 use ika_types::dwallet_mpc_error::{DwalletMPCError, DwalletMPCResult};
-use ika_types::messages_dwallet_mpc::EncryptedShareVerificationRequestEvent;
-use rand_core::OsRng;
+use ika_types::messages_dwallet_mpc::{
+    DWalletSessionEvent, EncryptedShareVerificationRequestEvent, MPCProtocolInitData, SessionInfo,
+};
 use twopc_mpc::dkg::Protocol;
 use twopc_mpc::secp256k1::class_groups::AsyncProtocol;
+
+pub(crate) fn start_encrypted_share_verification_session_info(
+    deserialized_event: DWalletSessionEvent<EncryptedShareVerificationRequestEvent>,
+) -> SessionInfo {
+    SessionInfo {
+        session_type: deserialized_event.session_type.clone(),
+        session_identifier: deserialized_event.session_identifier_digest(),
+        epoch: deserialized_event.epoch,
+        mpc_round: MPCProtocolInitData::EncryptedShareVerification(deserialized_event),
+    }
+}
 
 /// Verifies that the given encrypted secret key share matches the encryption of the dWallet's
 /// secret share, validates the signature on the dWallet's public share,
@@ -44,7 +57,7 @@ fn verify_centralized_secret_key_share_proof(
                 bcs::from_bytes(&dkg_public_output)?,
                 bcs::from_bytes(encryption_key)?,
                 bcs::from_bytes(encrypted_centralized_secret_share_and_proof)?,
-                &mut OsRng,
+                &mut OsCsRng,
             )
             .map_err(Into::<anyhow::Error>::into)?;
             Ok(())
