+++ crates/ika-protocol-config/src/lib.rs
@@ -262,7 +262,7 @@ pub struct ProtocolConfig {
     /// Configures the garbage collection depth for consensus. When is unset or `0` then the garbage collection
     /// is disabled.
     consensus_gc_depth: Option<u32>,
-    decryption_key_reshare_third_round_delay: Option<u32>,
+    decryption_key_reconfiguration_third_round_delay: Option<u32>,
     network_dkg_third_round_delay: Option<u32>,
 }
 
@@ -496,7 +496,7 @@ impl ProtocolConfig {
             consensus_max_num_transactions_in_block: Some(512),
             consensus_gc_depth: Some(60),
             // The delay is measured in consensus rounds.
-            decryption_key_reshare_third_round_delay: Some(10),
+            decryption_key_reconfiguration_third_round_delay: Some(10),
             network_dkg_third_round_delay: Some(10),
         };
