+++ crates/ika-swarm-config/Cargo.toml
@@ -15,7 +15,7 @@ anemo.workspace = true
 anyhow.workspace = true
 bcs.workspace = true
 fastcrypto.workspace = true
-rand.workspace = true
+rand = "0.8" # use old version for compatability with fastcrypto
 serde.workspace = true
 serde_with.workspace = true
 serde_yaml.workspace = true
