+++ crates/ika-types/Cargo.toml
@@ -19,7 +19,7 @@ byteorder.workspace = true
 consensus-config.workspace = true
 once_cell.workspace = true
 prometheus.workspace = true
-rand.workspace = true
+rand = "0.8" # use old version for compatability with fastcrypto
 serde.workspace = true
 serde-name.workspace = true
 thiserror.workspace = true
@@ -55,7 +55,7 @@ sui-config.workspace = true
 
 sui-types.workspace = true
 serde_repr.workspace = true
-rand_chacha = "0.3.1"
+rand_chacha = "0.3" # use old version for compatability with fastcrypto
 base64.workspace = true
 
 [dev-dependencies]
