+++ crates/message-digest/Cargo.toml
@@ -1,12 +1,9 @@
 [package]
-name = "shared_wasm_class_groups"
+name = "message-digest"
 edition = "2021"
 version.workspace = true
 
 [dependencies]
-twopc_mpc.workspace = true
-class_groups.workspace = true
-mpc.workspace = true
 group.workspace = true
 bcs = "0.1.6"
 base64.workspace = true
@@ -15,8 +12,8 @@ schemars.workspace = true
 anyhow.workspace = true
 dwallet-mpc-types.workspace = true
 k256.workspace = true
-sha2 = "0.10.8"
-sha3 = "0.10.8"
+sha2 = "0.11.0-rc.0"
+sha3 = "0.11.0-rc.0"
 
 [lints]
 workspace = true
