+++ scripts/update_graphql_e2e_test_protocol_version.py
