+++ sdk/dwallet-mpc-wasm/Cargo.toml
@@ -51,8 +51,10 @@ publish = false
 #]
 
 [dependencies]
-dwallet-mpc-centralized-party = { path = "../../crates/dwallet-mpc-centralized-party" }
+dwallet-mpc-centralized-party = { path = "../../crates/dwallet-mpc-centralized-party", features = ["wasm_js"] }
 anyhow = "1.0.95"
+getrandom = { version = "0.2.16", features = ["js"] } # TODO: idk why, but we need this, also this is old version
+group = { git = "https://github.com/dwallet-labs/cryptography-private", features = ["wasm_js"], rev = "7dd2b76"}
 
 wasm-bindgen = "0.2.100"
 serde-wasm-bindgen = "0.6.5"
@@ -61,5 +63,9 @@ serde-wasm-bindgen = "0.6.5"
 crate-type = ["cdylib", "rlib"]
 
 [patch.crates-io]
-ecdsa = { git = "https://github.com/yaelAbergel/signatures", rev = "7c8523182b1a28e5bcc6bb8d1508fbde1ea845c4"}
-crypto-bigint = { git = 'https://github.com/erik-3milabs/crypto-bigint.git', rev = 'a4da280' }
+sha2 = { git = 'https://github.com/RustCrypto/hashes.git', tag = "sha2-v0.11.0-rc.0" }
+sha3 = { git = 'https://github.com/RustCrypto/hashes.git', tag = "sha3-v0.11.0-rc.0" }
+digest = { git = 'https://github.com/RustCrypto/traits.git', tag = "digest-v0.11.0-rc.0" }
+signature = { git = 'https://github.com/RustCrypto/traits.git', tag = "signature-v3.0.0-rc.1" }
+curve25519-dalek = { git = 'https://github.com/dalek-cryptography/curve25519-dalek', rev = "b7e4fc0" }
+crypto-bigint = { git = 'https://github.com/erik-3milabs/crypto-bigint.git', rev = "640aacb" }
