+++ sdk/typescript/package.json
@@ -150,8 +150,8 @@
 		"bech32": "^2.0.0",
 		"gql.tada": "^1.8.10",
 		"graphql": "^16.11.0",
+		"js-sha3": "^0.9.3",
 		"tweetnacl": "^1.0.3",
-		"valibot": "^1.1.0",
-		"js-sha3": "^0.9.3"
+		"valibot": "^1.1.0"
 	}
 }
